moved multi-plane functionality in separate folder
